Activation over BSS networks
The activation code will now seek and connect to unsecured BSS networks,
and look for a lease server over IPv4 and IPv6. The addresses checked
represent the default configuration of the school server.

This may also fix the IPv4-over-mesh activation scenario.

BSS is checked before mesh. One final change is that if a lease server
is found but returns no lease (or an invalid lease), further networks/mesh
channels will be checked anyway.