Introduce greplease, and use it to read on-disk lease files
Lease files provided on disk can be huge, as we have seen in
large deployments (at their repair centers). Reading and parsing
those in-memory is extraordinarily expensive. Rough tests show
RAM use for a parsed leases file to be about 5x its size.

So for 'format 1' JSON lease files we use the greplease.grep()
function that searches for the key->val combo inthe leases file
and returns the value -- which is later handled by the 'singleton
lease' codeflow in find_leases()

If the lease is a singleton lease, then we read it straight away.

The greplease code uses mmap (so requires modulemmap.so).

An older version that tests for mmap and falls back to a read()-based
grep routine exists at
http://dev.laptop.org/git/users/martin/olpcrd-rootskel.git/