Stop using mtdblock driver (#11234)
Should have read the FAQ first. This is a bad idea:
http://www.linux-mtd.infradead.org/faq/general.html#L_ext2_mtd

It causes a drastic boot slowdown if block 0 is bad: udev (or whatever)
tries to read it during boot, causing a mountain of errors in the kernel
log.

Switch (back) to using the standard mtd driver, adding a olpc-mtd module
to support this. mtd support has already been contributed to dracut
upstream.