act-gui: fix framebuffer dimensions parsing
This code was using virtual X/Y resolution. On the XO-1.75, the
virtual Y resolution is 2402.

Move to using the visible resolution. Technically we should be
offsetting this by the xoffset/yoffset values, but these are 0
on all 3 laptop models.

Fixes act-gui graphics being rendered outside the visible area
on XO-1.75.