Use DHCP to connect to networks for activation
Instead of generating a random IPv4 address and hardcoding XS network
details, use DHCP for automatic network configuration.

Add the relevant components for DNS support, and additionally (primarily)
try to connect to the XS on its hostname 'schoolserver'

This allows for initramfs operation on networks that are not XS-owned
(e.g. with a schoolserver available, but perhaps a different addressing
scheme)