olpc-boot: increase root partition size (#10040)
When booting from partitioned SD cards, resize the system partition
to fill the disk.

We must do this from the initramfs, as doing it when the partition
is mounted means that the new sizes don't take effect until reboot.

We only resize when there is space for the partition to grow, and when
a "no-resize" marker file is not present.

The filesystem is not changed here, just the partition table.
The filesystem will be resized online by early-boot scripts.