olpc-boot-cmdline: initial support for DT bootpath strings
The move to devicetree changes the bootpath.

This creates an unfortunate flag day for XO-1.75; we'll figure out
a migration strategy shortly.

The device nodes selected from these addresses might be XO-4 specific,
need to check.

The way forward will be to find the appropriate device node in /sys
based on the path from the DT.