Pass /bootpart mount to real system
The initramfs has to mount the boot partition, and now that
the code to find the boot partition is becoming even more
complex, we might as well only have it in one place.

Move the /bootpart mount to the real system once we're done with it.
The equivalent /bootpart mounting code has been removed from olpc-utils.