Disable emergency shell by default
The .die indicator is no longer the valid way of denying a shell; that
just stops the initramfs from continuing when hitting an error (after
having provided a shell).

After this change, emergency shell is disabled by default, but
developers can add "rd.shell" to the kernel boot params to get it
back (this is explained in the error message too).