olpc-boot: add XO-1 external SD card bootpath
On the XO-1 the external SD card has not been supported for the
operating system because it is occasionally unreliable; bit errors on
some units.  The XO-1 units in the field are now unreliable for other
reasons, so this restriction can be removed.

Trying to use it exposed a problem; the root filesystem could not be
discovered, because there was no match for the bootpath given by Open
Firmware when XO-1 is booted from external SD card.

This patch adds recognition for the bootpath, and assumes that the SD
card has a root and boot partition in the same way as used on XO-1.5 and
later.

See http://dev.laptop.org/ticket/1220
See http://dev.laptop.org/ticket/11233
See http://dev.laptop.org/ticket/11872