olpc-boot: use DT only to get boot path on new Arm kernels
On pure DT based platforms, sysfs has enough information to make us
resolve the bootpath to a mmcblk device unambigiously. Use it on 5.0+ on
Arm.

In that case also avoid including the old cmdline hook; OFW recognizes
it and turns on compat bits that mess things up (lies about the boot
path and turns of FDT).